Fixes in existing tests (testCharacterGenerator)
